Add support for Splunk v8 and Python v3
We adapt the code so that it works exclusively with Python v3 (3.7
or greater).
The `lib/` folder was added which currently holds the version number,
but can be later used to easier extend the app.

The `setup.xml` has been extended with more inline help so that the
user installing the add-on has bit more information as to what to fill
in.

The `setup.py` was added with a custom `distutils` command called
`splunkdist`. This makes it easy to create a distribution ready to
be used in Splunk.

The README has been revamped with new documentation and new structure.

Resolves #18